Initialize Bitcoind after starting it
This should always go together anyway!
