+++ api_tests/lib/config.ts
@@ -1,7 +1,7 @@
 import * as tmp from "tmp";
 import { LedgerConfig } from "./ledgers/ledger_runner";
-import { BitcoinNodeConfig } from "./ledgers/bitcoin";
 import { EthereumNodeConfig } from "./ledgers/ethereum";
+import { BitcoinNodeConfig } from "./ledgers/bitcoind_instance";
 
 export interface CndConfigFile {
     http_api: HttpApi;
